% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{countBarcodes}
\alias{countBarcodes}
\title{Extract putative DNA molecules from barcoded BAM file.}
\usage{
countBarcodes(inp_fn, out_fn, wl_fn)
}
\arguments{
\item{inp_file}{input FQ file}

\item{out_file}{output counts file}

\item{out_file}{output counts file}
}
\description{
Extract putative DNA molecules from barcoded BAM file.
}
