% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode.R
\name{extractMolecules2}
\alias{extractMolecules2}
\title{Extract Molecules from barcoded BAM file}
\usage{
extractMolecules2(bam)
}
\arguments{
\item{bam}{bam filename}
}
\value{
data.frame
}
\description{
Extract Molecules from barcoded BAM file
}
